+++ src/baselime.ts
@@ -53,7 +53,6 @@ export class BaselimeSDK {
             resource: detectResourcesSync({
                 detectors: [awsEcsDetector, awsEc2Detector, awsLambdaDetector, new VercelDetector(), new KoyebDetector(), new ServiceDetector({ serviceName: this.options.service })],
             }),
-            forceFlushTimeoutMillis: 500,
         });
 
 
@@ -99,7 +98,7 @@ export class BaselimeSDK {
         }
         
         provider.register();
-
+        
         registerInstrumentations({
             instrumentations: [
                 ...this.options.instrumentations || []