+++ src/index.ts
@@ -4,3 +4,4 @@ export { BetterHttpInstrumentation } from './http.ts';
 export { StripePlugin } from './http-plugins/stripe.ts';
 export { HttpPlugin } from './http-plugins/plugin.ts';
 export { VercelPlugin } from './http-plugins/vercel.ts';
+export { wrap as withOpenTelemetry } from './lambda.ts';